implemented OpenCV to preprocess images
OpenCV's preprocessing is sued to grayscale, dilate, and erode images before being upload to the OCR so that we can get a more accurate image-to-text result by GoogleVision. It's been noticed that this algorithm has actually reduced GoogleVision's performance on Turkish text images, so until improved further, the algorithm is commented out.