major improvements to the OCR algorithm
instead of checking whether the second word's centerY is within the first word's topY and bottomY, the algorithm now kind of iteratively works with a new approach. First, it sorts the words by y-coordinates. Then, it gets all words that are in the same line (by checking if the next word's topY is above all the other word's most bottomY with a margin of error of range Y / 3) and then sorts these words by x-coordinates. Finally, it adds this line of words to the ultimate string. Then, it proceeds with the next line and so on. Works like a charm with the previous 8 examples.